refactor: out of bread item render
