+++ components/breadcrumb/Breadcrumb.tsx
@@ -2,14 +2,15 @@ import classNames from 'classnames';
 import toArray from 'rc-util/lib/Children/toArray';
 import pickAttrs from 'rc-util/lib/pickAttrs';
 import * as React from 'react';
-import { ConfigContext } from '../config-provider';
 import { cloneElement } from '../_util/reactNode';
 import warning from '../_util/warning';
+import { ConfigContext } from '../config-provider';
 import type { BreadcrumbItemProps } from './BreadcrumbItem';
 import BreadcrumbItem from './BreadcrumbItem';
 import BreadcrumbSeparator from './BreadcrumbSeparator';
 
 import useStyle from './style';
+import useItemRender from './useItemRender';
 import useItems from './useItems';
 
 /** @deprecated New of Breadcrumb using `items` instead of `routes` */
@@ -43,7 +44,7 @@ export interface BreadcrumbSeparatorType {
 
 export type ItemType = BreadcrumbItemType | BreadcrumbSeparatorType;
 
-type InternalRouteType = Partial<BreadcrumbItemType & BreadcrumbSeparatorType>;
+export type InternalRouteType = Partial<BreadcrumbItemType & BreadcrumbSeparatorType>;
 
 export interface BaseBreadcrumbProps {
   prefixCls?: string;
@@ -74,19 +75,6 @@ export interface NewBreadcrumbProps extends BaseBreadcrumbProps {
 }
 
 export type BreadcrumbProps = BaseBreadcrumbProps | LegacyBreadcrumbProps | NewBreadcrumbProps;
-
-function getBreadcrumbName(route: InternalRouteType, params: any) {
-  if (route.title === undefined) {
-    return null;
-  }
-  const paramsKeys = Object.keys(params).join('|');
-  return typeof route.title === 'object'
-    ? route.title
-    : String(route.title).replace(
-        new RegExp(`:(${paramsKeys})`, 'g'),
-        (replacement, key) => params[key] || replacement,
-      );
-}
 
 const getPath = (params: any, path?: string) => {
   if (path === undefined) {
@@ -132,13 +120,7 @@ const Breadcrumb: CompoundedComponent = (props) => {
     warning(!legacyRoutes, 'Breadcrumb', '`routes` is deprecated. Please use `items` instead.');
   }
 
-  const mergedItemRender =
-    itemRender ||
-    ((route: BreadcrumbItemType) => {
-      const name = getBreadcrumbName(route, params);
-
-      return name;
-    });
+  const mergedItemRender = useItemRender(itemRender);
 
   if (mergedItems && mergedItems.length > 0) {
     // generated by route
